Adding assets to ToParkBuildingHelper.cs
Related to Issue: Add more unique buildings to list #1.
I added the Temple, Pyramid of Safety, and the Statue of Colossalus. The place I was pulling the exact names of the assets did not list any of the other DLC's, so I'm not 100% sure what the names are. I've left them commented out on the pull request.